Extract pending work queue from logger into its own module
pending-work.ts is request-scoped infrastructure, not logging.
Logger and webhook now both import from #lib/pending-work.ts.
