+++ src/lib/webhook.ts
@@ -7,7 +7,8 @@ import { compact, map, unique } from "#fp";
 import { getAllowedDomain } from "#lib/config.ts";
 import { logActivity } from "#lib/db/activityLog.ts";
 import { getEnv } from "#lib/env.ts";
-import { ErrorCode, addPendingWork, logError } from "#lib/logger.ts";
+import { ErrorCode, logError } from "#lib/logger.ts";
+import { addPendingWork } from "#lib/pending-work.ts";
 import type { ContactInfo } from "#lib/types.ts";
 import { nowIso } from "#lib/now.ts";
 import { getBusinessEmailFromDb } from "#lib/business-email.ts";
