+++ src/routes/index.ts
@@ -8,7 +8,8 @@ import { isSetupComplete } from "#lib/config.ts";
 import { loadCurrencyCode } from "#lib/currency.ts";
 import { loadTheme } from "#lib/theme.ts";
 import { runWithQueryLogContext } from "#lib/db/query-log.ts";
-import { createRequestTimer, ErrorCode, flushPendingWork, logError, logRequest, runWithRequestId } from "#lib/logger.ts";
+import { createRequestTimer, ErrorCode, logError, logRequest, runWithRequestId } from "#lib/logger.ts";
+import { flushPendingWork } from "#lib/pending-work.ts";
 import {
   applySecurityHeaders,
   contentTypeRejectionResponse,
